Move processor containers to parent folder
