+++ src/renderer/containers/HighlightOptions.tsx
