+++ src/renderer/containers/ProcessorStatusBar.tsx
