+++ src/renderer/views/main/ReplayProcessorView.tsx
@@ -9,7 +9,7 @@ import { useSelector } from "react-redux";
 
 import { Text } from "@/components/Form";
 import { ProgressBar } from "@/components/ProgressBar";
-import { ProcessorStatusBar } from "@/containers/processor/ProcessorStatusBar";
+import { ProcessorStatusBar } from "@/containers/ProcessorStatusBar";
 import { ComboFinder } from "@/containers/ComboFinder";
 
 const Content = styled.div`
